+++ src/main/java/com/ryu/studyhelper/auth/token/RedisRefreshTokenService.java
@@ -2,7 +2,7 @@ package com.ryu.studyhelper.auth.token;
 
 import com.ryu.studyhelper.common.enums.CustomResponseStatus;
 import com.ryu.studyhelper.common.exception.CustomException;
-import com.ryu.studyhelper.config.jwt.util.JwtUtil;
+import com.ryu.studyhelper.config.security.jwt.JwtUtil;
 import com.ryu.studyhelper.member.repository.MemberRepository;
 import com.ryu.studyhelper.member.domain.Member;
 import lombok.RequiredArgsConstructor;
